shopping cart final functionality

+++ MyBookStore/src/main/java/com/group10/book/BookRepository.java
@@ -2,6 +2,6 @@ package com.group10.book;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookRepository extends CrudRepository<Book, Integer> {
-    public Long countById(Integer id);
+public interface BookRepository extends CrudRepository<Book, Long> {
+    public Long countById(Long id);
 }
